fix: field validation
