+++ src/constants/types.ts
@@ -8,13 +8,6 @@ export interface CNavLinkProps {
   activeIcon: JSX.Element | React.ReactNode;
 }
 
-export type FormValues = {
-  store: string;
-  email: string;
-  password: string;
-  confirmPassword: string;
-};
-
 export type SvgProps = {
   fill?: string;
 };
