+++ src/containers/SignupContainer/form.tsx
@@ -5,30 +5,34 @@ import { useRouter } from 'next/navigation';
 import { Form, Field } from 'react-final-form';
 
 import { Typography } from '@/assets';
-
 import Pages from '@/constants/pages';
-import { FormValues } from '@/constants/types';
-
 import CInput from '@/components/CInput';
 import CButton from '@/components/CButton';
 import CCheckbox from '@/components/CCheckbox';
-
 import { required, minLength, validateEmail, validatePassword } from '@/utils/validators';
 
-const SignUpForm = () => {
-  const [isRememberChecked, setIsRememberChecked] = useState(false);
+export type FormValues = {
+  store: string;
+  email: string;
+  password: string;
+  confirmPassword: string;
+};
 
+const SignUpForm = () => {
   const routes = useRouter();
+  const [isRememberChecked, setIsRememberChecked] = useState(false);
 
   const handleRedirectToSignIn = () => {
     routes.push(Pages.SIGNIN);
   };
 
-  const handleRememberChange = () => {
-    setIsRememberChecked(!isRememberChecked);
+  const handleRememberChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setIsRememberChecked(e.target.checked);
   };
 
   const onSubmit = (values: FormValues) => {
+    console.log(isRememberChecked);
+
     console.log(values);
   };
 
@@ -52,26 +56,28 @@ const SignUpForm = () => {
       <div className="flex justify-start items-start">
         <Typography />
       </div>
+
       <div className="flex justify-center mobile:items-start mobile:mt-8 items-center h-full">
         <div className="flex flex-col w-full">
           <div className="my-4 space-y-4">
             <p className="text-2xl font-medium text-darkGreen select-none">Sign Up</p>
           </div>
+
           <Form
             onSubmit={onSubmit}
             validate={validateForm}
-            render={({ handleSubmit, submitError, submitting, pristine }) => (
+            render={({ handleSubmit, submitError, submitting, pristine, invalid }) => (
               <form onSubmit={handleSubmit} autoComplete="false">
                 <div className="space-y-3 w-full">
                   <Field
                     name="store"
-                    validate={composeValidators(required, minLength(8))}
+                    validate={composeValidators(required, minLength(4))}
                     render={({ input, meta }) => (
                       <CInput
+                        {...input}
                         border
                         autoFocus
-                        placeholder="store name"
-                        input={input}
+                        placeholder="Store Name"
                         meta={meta}
                         error={meta.touched && meta.error}
                         errorMsg={meta.error}
@@ -80,13 +86,13 @@ const SignUpForm = () => {
                   />
 
                   <Field
-                    name="Email"
+                    name="email"
                     validate={composeValidators(required, validateEmail)}
                     render={({ input, meta }) => (
                       <CInput
+                        {...input}
                         border
                         placeholder="Email"
-                        input={input}
                         meta={meta}
                         error={meta.touched && meta.error}
                         errorMsg={meta.error}
@@ -99,10 +105,10 @@ const SignUpForm = () => {
                     validate={composeValidators(required, minLength(8), validatePassword)}
                     render={({ input, meta }) => (
                       <CInput
+                        {...input}
                         border
                         type="password"
                         placeholder="Password"
-                        input={input}
                         meta={meta}
                         error={meta.touched && meta.error}
                         errorMsg={meta.error}
@@ -115,10 +121,10 @@ const SignUpForm = () => {
                     validate={composeValidators(required, minLength(8))}
                     render={({ input, meta }) => (
                       <CInput
+                        {...input}
                         border
                         type="password"
                         placeholder="Confirm Password"
-                        input={input}
                         meta={meta}
                         error={meta.touched && meta.error}
                         errorMsg={meta.error}
@@ -126,18 +132,17 @@ const SignUpForm = () => {
                     )}
                   />
                 </div>
+
                 <div className="w-full mt-2 space-y-3">
                   <Field
                     name="remember"
-                    render={({ input, meta }) => (
+                    render={() => (
                       <CCheckbox
                         label={<p className="!text-smokyBlue">Remember me</p>}
                         checked={isRememberChecked}
                         onChange={handleRememberChange}
                         value="remember"
                         className="mt-3 -ml-[6px]"
-                        input={input}
-                        meta={meta}
                       />
                     )}
                   />
@@ -148,7 +153,7 @@ const SignUpForm = () => {
                     text="Sign Up"
                     className="mt-4"
                     type="submit"
-                    disabled={pristine || submitting}
+                    disabled={pristine || submitting || invalid}
                   />
                 </div>
               </form>
@@ -156,8 +161,10 @@ const SignUpForm = () => {
           />
         </div>
       </div>
+
       <div className="flex justify-between items-center w-full">
         <p className="text-sm select-none text-smokyBlue">Already have an Account?</p>
+
         <CButton
           text="Sign in"
           variant="outline"
